backend/requirements.txt
alembic>=1.13.1
asyncpg>=0.29.0
fastapi>=0.104.0
pydantic>=2.4.0
pydantic-core>=2.10.0
sqlalchemy[asyncio]>=2.0.29
uvicorn>=0.29.0
psycopg2-binary>=2.9.9
# OCR/NLP dependencies
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
spacy>=3.7.0
ru-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.7.0/ru_core_news_lg-3.7.0-py3-none-any.whl
# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-mock>=3.10.0
# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
